Refactored WeatherDTO to use int instead of long

+++ src/main/java/ee/taltech/testing/courseproject/DTO/WeatherDTO.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class WeatherDTO {
-    private long temp;
-    private long pressure;
-    private long humidity;
+    private int temp;
+    private int pressure;
+    private int humidity;
 }
